+++ NEWS.md
@@ -1,5 +1,7 @@
 ## Unreleased changes
 - Added support for VHDL-2008 type generics on packages and entities.
+- New command line options `-P` and `-G` can be used to set the size of
+  the process and global secondary stack respectively.
 
 ## Version 1.6.1 - 2022-02-05
 - Fix compatibility with BSD make (#440).